vrops-inventory/requirements.txt
requests>=2.25.1
pyyaml>=5.4.1
ansible-core>=2.11.0